requirements.txt
asgiref==3.2.3
certifi==2019.11.28
dj-database-url==0.5.0
Django==3.0.2
django-json-widget==0.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
numpy==1.18.1
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
wincertstore==0.2
django-admin-rangefilter